requirements.txt
Django>=4.2, <4.3
pillow==11.1.0
python-decouple==3.8
django-sequences==3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djangorestframework-simplejwt
PyJWT==2.9.0
cryptography==44.0.2
drf-yasg==1.21.9
celery==5.4.0
requests==2.32.3
django-filter==25.1
django-cors-headers==4.7.0
django-celery-beat==2.7.0
psycopg2-binary==2.9.10
django-redis==5.4.0
simplejson==3.20.1
django-celery-results==2.5.1
djangoql==0.18.1
django-json-widget==2.0.1
django-cryptography==1.1
coverage==7.6.12
flake8==7.1.2
aiohttp==3.10.11
aiohttp-retry==2.8.3
certifi==2024.7.4
python-dateutil==2.9.0.post0
